load dataframes

+++ src/scrapers/data_cleanup/koronawirusunas.py
@@ -1,5 +1,6 @@
 import os
 from pathlib import Path
+import pandas as pd
 
 
 def clean_data():
@@ -11,7 +12,12 @@ def clean_data():
     # Names of files to be deleted
     files_to_rm = ["data", "Data_przyrost_procent",
                    "datah", "datazgony", "datazk",
-                   "datazz", "populationData"]
+                   "datazz", "populationData",
+                   "Data_przyrost_szpital", "Data_przyrost_testy"]
+
+    # Names of files to be cleaned
+    files_to_clean = ["dataSource_hospitalizacja", "dataSource_mobilnosc",
+                      "dataSource_przyrost", "dataSource_testy"]
 
     # Get path to /src/scrapers/Data/koronawirusunas
     path = Path(__file__).parents[1]
@@ -20,9 +26,54 @@ def clean_data():
     # Remove files
     for filename in files_to_rm:
         filepath = Path(path, filename + ".csv")
-        os.remove(filepath)
-        print("Removed file: {}".format(filepath))
+        if os.path.isfile(filepath):
+            os.remove(filepath)
+            print("Removed file: {}".format(filepath))
 
+    # Check if files to clean exist
+    for filename in files_to_clean:
+        filepath = Path(path, filename + ".csv")
+        if os.path.isfile(filepath):
+            print("Found file: {}".format(filepath))
+        else:
+            # Exit if file is not present
+            print("Error missing file: {}".format(filepath))
+            return
+
+    # Load dataframes
+    hospitalizacja = pd.read_csv(Path(path, files_to_clean[0] + ".csv"),
+                                 sep="\t",
+                                 index_col=0,
+                                 usecols=["country", "hosp", "kwar", "kwar_z", "nadzor"]
+                                 )
+
+    mobilnosc = pd.read_csv(Path(path, files_to_clean[1] + ".csv"),
+                            sep="\t",
+                            index_col=0,
+                            usecols=["dzien", "pieszo", "pojazdem"]
+                            )
+
+    przyrost = pd.read_csv(Path(path, files_to_clean[2] + ".csv"),
+                           sep="\t",
+                           index_col=0,
+                           usecols=["country", "zar", "chor", "zgo", "wyl"]
+                           )
+    testy = pd.read_csv(Path(path, files_to_clean[3] + ".csv"),
+                        sep="\t",
+                        index_col=0,
+                        usecols=["dzien", "smp", "testy", "testyl"]
+                        )
+
+    # Remove index.name
+    hospitalizacja.index.name = None
+    mobilnosc.index.name = None
+    przyrost.index.name = None
+    testy.index.name = None
+
+    print(hospitalizacja)
+    print(mobilnosc)
+    print(przyrost)
+    print(testy)
 
 if __name__ == "__main__":
     clean_data()
